chore: moved ContributorList into it's own file
