+++ pages/lists/[listId]/contributors.tsx
@@ -1,56 +1,10 @@
 import { GetServerSidePropsContext } from "next";
 import { createPagesServerClient } from "@supabase/auth-helpers-nextjs";
 import ListPageLayout from "layouts/lists";
-import ContributorTable from "components/organisms/ContributorsTable/contributors-table";
-import PaginationResults from "components/molecules/PaginationResults/pagination-result";
-import ContributorListTableHeaders from "components/molecules/ContributorListTableHeader/contributor-list-table-header";
-import Pagination from "components/molecules/Pagination/pagination";
 import { fetchApiData, validateListPath } from "helpers/fetchApiData";
 import Error from "components/atoms/Error/Error";
-import useSupabaseAuth from "lib/hooks/useSupabaseAuth";
 import { convertToContributors, useContributorsList } from "lib/hooks/api/useContributorList";
-import ClientOnly from "components/atoms/ClientOnly/client-only";
-
-interface ContributorsListProps {
-  contributors: DbPRContributor[];
-  isLoading: boolean;
-  meta: Meta;
-  setPage: (page: number) => void;
-}
-
-const ContributorsList = ({ contributors, isLoading, meta, setPage }: ContributorsListProps) => {
-  return (
-    <div className="lg:min-w-[1150px] px-16 py-8">
-      <ContributorListTableHeaders />
-      <ClientOnly>
-        <ContributorTable loading={isLoading} topic={"*"} contributors={contributors} />
-      </ClientOnly>
-      <div className="flex items-center justify-between w-full py-1 md:py-4 md:mt-5">
-        <div>
-          <div className="">
-            <PaginationResults metaInfo={meta} total={meta.itemCount} entity={"contributors"} />
-          </div>
-        </div>
-        <div>
-          <div className="flex flex-col gap-4">
-            <Pagination
-              pages={[]}
-              hasNextPage={meta.hasNextPage}
-              hasPreviousPage={meta.hasPreviousPage}
-              totalPage={meta.pageCount}
-              page={meta.page}
-              onPageChange={function (page: number): void {
-                setPage(page);
-              }}
-              divisor={true}
-              goToPage
-            />
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
+import ContributorsList from "components/organisms/ContributorsList/contributors-list";
 
 export const getServerSideProps = async (ctx: GetServerSidePropsContext) => {
   const supabase = createPagesServerClient(ctx);
@@ -97,10 +51,7 @@ interface ContributorListPageProps {
   isError: boolean;
 }
 
-type UserID = number | null | undefined;
-
 const ContributorsListPage = ({ list, initialData, isError }: ContributorListPageProps) => {
-  const { userId } = useSupabaseAuth();
   // create useIsOwner(list?.user_id, userId) once we're ready to implement this.
   const isOwner = false;
   const {
